# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "This is a sample API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/files/{file}/download": {
            "get": {
                "tags": [
                    "Payment Requests"
                ],
                "summary": "Download payment request attachment file",
                "operationId": "e10941755b161b1cc7cda6152c6067a7",
                "parameters": [
                    {
                        "name": "file",
                        "in": "path",
                        "description": "File ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "9de1f353-4208-4620-901a-dc1a5978c374"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File downloaded successfully",
                        "headers": {
                            "Content-Disposition": {
                                "description": "attachment; filename={filename}",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}